Move payload objects into their own file.
Moved payload objects into their own file, along with combining duplicate payloads.
