+++ src/models/mod.rs
@@ -2,4 +2,5 @@ pub mod authentication;
 pub mod document;
 pub mod error;
 pub mod paste;
+pub mod payload;
 pub mod snowflake;
